+++ 3rd-party-integrations/SentryCocoaLumberjack/Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:6.0
+// swift-tools-version:6.1
 import PackageDescription
 
 let package = Package(
@@ -10,16 +10,26 @@ let package = Package(
             targets: ["SentryCocoaLumberjack"]
         )
     ],
+    traits: [
+        .default(enabledTraits: ["SentryFromBinary"]),
+        .init(name: "SentryFromBinary", description: "Use precompiled Sentry binary xcframeworks."),
+        .init(name: "SentryFromSource", description: "Build Sentry from source instead of using precompiled binaries.")
+    ],
     dependencies: [
         .package(url: "https://github.com/CocoaLumberjack/CocoaLumberjack", from: "3.8.0"),
-        .package(url: "https://github.com/getsentry/sentry-cocoa", from: "9.25.0")
+        .package(url: "https://github.com/getsentry/sentry-cocoa", from: "9.26.0")
     ],
     targets: [
         .target(
             name: "SentryCocoaLumberjack",
             dependencies: [
                 .product(name: "CocoaLumberjackSwift", package: "CocoaLumberjack"),
-                .product(name: "Sentry", package: "sentry-cocoa")
+                .product(name: "Sentry", package: "sentry-cocoa", condition: .when(traits: ["SentryFromBinary"])),
+                .product(name: "SentrySPM", package: "sentry-cocoa", condition: .when(traits: ["SentryFromSource"]))
+            ],
+            swiftSettings: [
+                .define("SENTRY_FROM_BINARY", .when(traits: ["SentryFromBinary"])),
+                .define("SENTRY_FROM_SOURCE", .when(traits: ["SentryFromSource"]))
             ]
         ),
         .testTarget(
@@ -27,7 +37,12 @@ let package = Package(
             dependencies: [
                 "SentryCocoaLumberjack",
                 .product(name: "CocoaLumberjackSwift", package: "CocoaLumberjack"),
-                .product(name: "Sentry", package: "sentry-cocoa")
+                .product(name: "Sentry", package: "sentry-cocoa", condition: .when(traits: ["SentryFromBinary"])),
+                .product(name: "SentrySPM", package: "sentry-cocoa", condition: .when(traits: ["SentryFromSource"]))
+            ],
+            swiftSettings: [
+                .define("SENTRY_FROM_BINARY", .when(traits: ["SentryFromBinary"])),
+                .define("SENTRY_FROM_SOURCE", .when(traits: ["SentryFromSource"]))
             ]
         )
     ]