+++ 3rd-party-integrations/SentrySwiftyBeaver/Tests/SentryDestinationTests.swift
@@ -1,5 +1,9 @@
 @testable import SentrySwiftyBeaver
+#if SENTRY_FROM_SOURCE
+import SentrySwift
+#else
 import Sentry
+#endif
 import SwiftyBeaver
 import XCTest